Add NextScan
