+++ src/pt/nextscan/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Next Scan'
+    extClass = '.NextScan'
+    themePkg = 'etoshore'
+    baseUrl = 'https://nextscan.cloud'
+    overrideVersionCode = 0
+    isNsfw = true
+}
+
+apply from: "$rootDir/common.gradle"
